utils on another script
